+++ Bermuda/Bermuda/DrawableEntity.h
@@ -12,7 +12,7 @@ private:
 	Image* drawImage;
 	SDL_Rect* sizeRect;
 public:
-	DrawableEntity(int id, int x, int y, int chunkSize, Image* image);
+	DrawableEntity(int id, double x, double y, int chunkSize, Image* image);
 	void draw(Camera* camera, SDL_Renderer* renderer);
 	void setDrawImage(Image* image);
 	virtual ~DrawableEntity();
